etherlab: lock version and fix library compile
This fixes a build issue where .o files compiled for the kernel were
reused in the library. This would cause vfp errors, soft/hardfloat
errors and wchar errors depending on the compiler configuration. This
was due to the Linux kernel preventing the compiler from using the
mentioned features. Since the `master` directory was shared between the
kernel and library builds, `make` would reuse .o files that it
shouldn't.

This also makes some changes to make this package submittable to
Buildroot:

1. Lock down the version. Branches aren't allowed since they may change
over time.
2. Add a .hash file to catch download corruption of strange things
happening with the source code.
3. Align the ChangeLog creation to how other Buildroot packages do it.
Unfortunately, this removes a nice comment, but the Buildroot devs like
consistency and this makes it consistent.